Fix controllers tests and database beans
