+++ application/src/main/java/ru/voting/controllers/GetInfoController.java
@@ -1,4 +1,4 @@
-package ru.voting.handlers;
+package ru.voting.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -6,19 +6,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import ru.voting.common.Poll;
 import ru.voting.common.User;
 import ru.voting.storage.DatabaseService;
-
-import java.util.*;
 
 @Controller
 public class GetInfoController {
     @Autowired
-    private DatabaseService database;
+    private DatabaseService databaseService;
     @GetMapping("/get_info")
     public ResponseEntity<String> getInfo(@RequestParam String email) {
-        User user = database.getById(User.class, email);
+        User user = databaseService.getById(User.class, email);
 
         if (user == null) {
             return new ResponseEntity<>("User not found", HttpStatus.NOT_FOUND);
