+++ application/src/main/java/ru/voting/controllers/GetPollController.java
@@ -1,4 +1,4 @@
-package ru.voting.handlers;
+package ru.voting.controllers;
 
 import org.springframework.web.bind.annotation.RestController;
 import ru.voting.common.Poll;
@@ -10,18 +10,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import ru.voting.storage.DatabaseService;
 
-import java.util.Map;
-
 @RestController
 public class GetPollController {
     @Autowired
     private EmailService emailService;
     @Autowired
-    private DatabaseService database;
+    private DatabaseService databaseService;
 
     @GetMapping("/get_poll")
     public ResponseEntity<String> getPoll(@RequestParam String id) {
-        Poll poll = database.getById(Poll.class, id);
+        Poll poll = databaseService.getById(Poll.class, id);
 
         if (poll == null) {
             return new ResponseEntity<>("Id not found", HttpStatus.NOT_FOUND);
