+++ application/src/main/java/ru/voting/storage/DatabaseConfig.java
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import ru.voting.common.Participant;
 import ru.voting.common.Poll;
+import ru.voting.common.PollAnswer;
 import ru.voting.common.User;
 
 @Configuration
@@ -15,6 +16,7 @@ public class DatabaseConfig {
                 .addAnnotatedClass(Poll.class)
                 .addAnnotatedClass(User.class)
                 .addAnnotatedClass(Participant.class)
+                .addAnnotatedClass(PollAnswer.class)
                 .buildSessionFactory();
     }
 }
