+++ application/src/test/java/ru/voting/controllers/GetInfoTest.java
@@ -1,25 +1,19 @@
-package ru.voting.handlers;
+package ru.voting.controllers;
 
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.test.web.servlet.MockMvc;
 import ru.voting.common.Poll;
 import ru.voting.common.User;
+import ru.voting.storage.DatabaseService;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
-import static org.springframework.boot.context.properties.bind.Bindable.mapOf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -27,34 +21,46 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class GetPollTest {
+public class GetInfoTest {
     @Autowired
     MockMvc mockMvc;
 
     @MockBean
-    Map<String, Poll> polls;
+    DatabaseService database;
 
     @Test
-    void testGetPoll() throws Exception {
-        Poll poll = new Poll(
-                "unique_id",
-                new User(
-                        "email",
-                        "password"
-                ),
+    void testGetInfo() throws Exception {
+        User eleazar = new User(
+                "e@mail.ru",
+                "qwerty123",
+                null
+        );
+
+        Poll poll1 = new Poll(
+                "id_1",
+                eleazar.getEmail(),
                 "Be or not to be?",
-                Map.of("be", 0,
-                        "do not be", 0),
-                Arrays.asList("a@b.ru", "c@d.ru"),
+                null, null,
+                0
+        );
+        Poll poll2 = new Poll(
+                "id_2",
+                eleazar.getEmail(),
+                "Tea or coffey?",
+                null, null,
                 0
         );
 
-        when(polls.containsKey(eq("unique_id"))).thenReturn(true);
-        when(polls.get(eq("unique_id"))).thenReturn(poll);
+        eleazar.setPolls(Arrays.asList(poll1, poll2));
 
-        mockMvc.perform(get("/get_poll?id=unique_id"))
+        when(database.getById(User.class, "e@mail.ru")).thenReturn(eleazar);
+
+        mockMvc.perform(get("/get_info?email=e@mail.ru"))
                 .andDo(print())
                 .andExpect(status().isOk())
-                .andExpect(content().string(poll.toString()));
+                .andExpect(content().string(
+                        Arrays.asList(poll1, poll2).toString()
+                ));
     }
+
 }
